restructure pakages
