+++ src/main/java/uk/ashleybye/httpserver/http/RequestParser.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.httpserver.server;
+package uk.ashleybye.httpserver.http;
 
 public interface RequestParser {
