+++ src/main/java/uk/ashleybye/httpserver/http/Response.java
@@ -1,4 +1,4 @@
-package uk.ashleybye.httpserver.server;
+package uk.ashleybye.httpserver.http;
 
 public interface Response {
