+++ src/test/java/uk/ashleybye/httpserver/http/HttpRequestParserStub.java
@@ -1,8 +1,5 @@
 package uk.ashleybye.httpserver.http;
 
-import uk.ashleybye.httpserver.server.Request;
-import uk.ashleybye.httpserver.server.RequestParser;
-
 public class HttpRequestParserStub implements RequestParser {
 
   private final HttpRequestStub requestStub;
