+++ src/test/java/uk/ashleybye/httpserver/http/HttpResponseSerializerStub.java
@@ -1,8 +1,5 @@
 package uk.ashleybye.httpserver.http;
 
-import uk.ashleybye.httpserver.server.Response;
-import uk.ashleybye.httpserver.server.ResponseSerializer;
-
 public class HttpResponseSerializerStub implements ResponseSerializer {
 
   @Override
