+++ src/test/java/uk/ashleybye/httpserver/http/HttpResponseSerializerTest.java
@@ -3,7 +3,6 @@ package uk.ashleybye.httpserver.http;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
-import uk.ashleybye.httpserver.server.Response;
 
 class HttpResponseSerializerTest {
